Add AbstractCommandWaitOperatorFactory
